added server-json and initiate a routes organization
