+++ index.html
@@ -10,7 +10,7 @@
     <link href="https://fonts.cdnfonts.com/css/gotham" rel="stylesheet">
     <title>Spotistas</title>
   </head>
-  <body>
+  <body class="bg-bgMain">
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
