+++ src/App.tsx
@@ -1,10 +1,12 @@
-import { Home } from './components/Home'
+import { Footer } from './components/Footer'
+import { Home } from './pages/Home'
 import './styles/global.css'
 
 function App() {
   return (
     <div>
       <Home />
+      <Footer />
     </div>
   )
 }
